feat(ripple): allow customization of ripple styling
BREAKING CHANGE: the mdcRipple directive does not set the class
mdc-ripple-surface on it's element anymore. This allows for other
classes to be used with mdcRipple, that can be customized with the
provided sass-mixins. For the old behavior, add the `surface` property
to your mdcRipple. This will add the mdc-ripple-surface class as before.